Set up to automatically pop up Chrome extensions
Added a new parameter to the `DAppConnector` constructor that allows
defining the extension IDs that automatically pop up when pairing or
sending requests to the wallet.

This contribution includes additional functions that allow for checking
whether a specific extension is installed and responding. The available
extensions are listed under the `extensions` property of the
`DAppConnector` class. To establish a session that will automatically
activate the extension, use the `connectExtension` method instead of the
`connect` method.

Signed-off-by: Fran Fernandez <fran@kabila.app>